requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
playwright==1.41.2
python-dateutil==2.8.2
schedule==1.2.1
